+++ Resources/Locale/en-US/administration/commands/set-outfit-command.ftl
@@ -1,4 +1,4 @@
-set-outfit-command-description = Sets the outfit of the specified entity. The entity must have {INDEFINITE($requiredComponent)} {$requiredComponent}
-set-outfit-command-help-text = Usage: {$command} <entityUid> | {$command} <entityUid> <outfitId>
-set-outfit-command-is-not-player-error = This command requires both arguments to work from the server console. 
-set-outfit-command-invalid-outfit-id-error = Invalid outfit id
+cmd-setoutfit-desc = Sets the outfit of the specified entity. The entity must have {INDEFINITE($requiredComponent)} {$requiredComponent}
+cmd-setoutfit-help = Usage: setoutfit <entityUid> | setoutfit <entityUid> <outfitId>
+cmd-setoutfit-is-not-player-error = This command requires both arguments to work from the server console.
+cmd-setoutfit-invalid-outfit-id-error = Invalid outfit id.
